requirements.txt
urllib3==1.25.3
colorlog==6.7.0
pymongo==3.13.0
Flask==2.0.3
Werkzeug==2.0.3
celery>=5.2.0,<6.0.0
flask-restx==1.0.3
requests==2.26.0
dnspython==2.2.1
tld==0.12.6
PyYAML>=6.0
geoip2==2.9.0
pyOpenSSL==22.1.0
mmh3==3.0.0
pyquery==1.4.3
openpyxl==3.0.0
gunicorn==20.1.0
psutil==5.7.2
crontab==0.23.0
pyparsing==3.1.1
# 独立部署额外依赖
pika==1.3.2
redis==4.5.4